Feature/user management/login

+++ usermanagement/src/main/java/com/example/usermanagement/dto/UserDto.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 @Data
 public class UserDto {
-    private Long userId;
+    private Integer userId;
     private String username;
     private String email;
     private Date createdAt;
     private Date updatedAt;
-    private List<String> roles; // role names only
+    private List<String> roles;
 }
